pyproject.toml
[tool.poetry]
name = "fastapi-lab2"
version = "0.1.0"
description = ""
authors = ["Pupupupupipu <vika566384@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
postgres = "^4.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
psycopg = "^3.1.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
